requirements.txt
web3
requests
eth_utils
python-dotenv
scipy